requirements.txt
PyYAML>=6.0
requests>=2.28.0
pytest>=7.0.0

